namespace slack-app models so they do not mess up the global namespace so bad
