--- plugins/slack_app/app/models/deploy_response_url.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-class DeployResponseUrl < ActiveRecord::Base
-  belongs_to :deploy, inverse_of: false
-end
