+++ plugins/slack_app/app/models/samson_slack_app/deploy_response_url.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module SamsonSlackApp
+  class DeployResponseUrl < ActiveRecord::Base
+    belongs_to :deploy, inverse_of: false
+  end
+end
