--- plugins/slack_app/app/models/slack_identifier.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-class SlackIdentifier < ActiveRecord::Base
-  belongs_to :user, optional: true, inverse_of: false
-
-  def self.app_token
-    SlackIdentifier.find_by_user_id(nil)&.identifier
-  end
-end
