+++ plugins/slack_app/lib/samson_slack_app/samson_plugin.rb
@@ -4,7 +4,7 @@ module SamsonSlackApp
   end
 end
 
-callback = ->(deploy, _) { SlackMessage.new(deploy).deliver }
+callback = ->(deploy, _) { SamsonSlackApp::SlackMessage.new(deploy).deliver }
 
 Samson::Hooks.callback :before_deploy, &callback
 Samson::Hooks.callback :after_deploy, &callback
