--- plugins/slack_app/test/fixtures/slack_identifiers.yml
@@ -1,15 +1,0 @@
-deployer:
-  identifier: 'Udeployer'
-  user: deployer
-
-deployer_buddy:
-  identifier: 'Ubuddy'
-  user: deployer_buddy
-
-super_admin:
-  identifier: 'Uadmin'
-  user: super_admin
-
-github_viewer:
-  identifier: 'Ugithubviewer'
-  user: github_viewer
