--- plugins/slack_app/test/models/deploy_response_url_test.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-require_relative '../test_helper'
-
-SingleCov.covered!
-
-describe DeployResponseUrl do
-end
