+++ plugins/slack_app/test/models/samson_slack_app/deploy_response_url_test.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+require_relative '../../test_helper'
+
+SingleCov.covered!
+
+describe SamsonSlackApp::DeployResponseUrl do
+end
