+++ plugins/slack_app/test/models/samson_slack_app/slack_identifier_test.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+require_relative '../../test_helper'
+
+SingleCov.covered! uncovered: 1
+
+describe SamsonSlackApp::SlackIdentifier do
+end
