--- plugins/slack_app/test/models/slack_identifier_test.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-require_relative '../test_helper'
-
-SingleCov.covered! uncovered: 1
-
-describe SlackIdentifier do
-end
